version display improvements: - use argparse action='version' to display version and exit - webrecorder-player has at least 1 required warc param - move all version-building related to versionbuild.py, separate from assetutils.py - slightly faster version load in standalone app
